requirements.txt
absl-py==1.0.0
astunparse==1.6.3
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
Flask==2.0.2
flatbuffers==2.0
gast==0.4.0
gensim==3.8.3
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
h5py==3.6.0
idna==3.3
importlib-metadata==4.10.0
itsdangerous==2.0.1
Jinja2==3.0.3
JPype1==1.3.0
keras==2.7.0
Keras-Preprocessing==1.1.2
konlpy==0.6.0
libclang==12.0.0
lxml==4.7.1
Markdown==3.3.6
MarkupSafe==2.0.1
numpy==1.22.0
oauthlib==3.1.1
opt-einsum==3.3.0
protobuf==3.19.1
py-hanspell @ git+https://github.com/ssut/py-hanspell.git@8e993cf46f97f9d665c15633a0fc78ac1b727713
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scipy==1.7.3
six==1.16.0
smart-open==5.2.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
typing_extensions==4.0.1
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.7.0
